change sum data type?

+++ function-1-2.cpp
@@ -3,12 +3,12 @@ double array_mean(int array[], int n) {
         return 0.0;
     }
 
-    int sum = 0;
+    double sum = 0;
     for (int i = 0; i < n; i++){
         sum += array[i];
     }
 
-    double average = double(sum / n);
+    double average = sum / n;
 
     return average;
